requirements.txt
Flask>=3.0.3
gunicorn>=23.0.0
uvicorn>=0.34.2
lxml>=5.2.1
PyMySQL>=1.1.1
mysql-connector-python>=9.3.0
mysqlclient>=2.2.7
matplotlib>=3.9.2
jsonpointer>=2.1
protobuf>=4.25.3
psycopg==3.2.6
redis>=5.0.0